[front] - (integrationFeed) use existing component to oneclick deploy #11297
